Moved Reddit Feed and Discord import into separate module
Now possible to run the Reddit feed separately from the main bot, which
handles command responses.
